@@ -2,7 +2,11 @@
 
   <PropertyGroup>
     <OutputType>Exe</OutputType>
-    <TargetFramework>netcoreapp3.1</TargetFramework>
+      <TargetFramework>net5.0</TargetFramework>
+    <!--
+      To target Windows IoT, use the following TargetFramework property instead, or multi-target
+      <TargetFramework>net5.0-windows10.0.17763.0</TargetFramework>
+    --> 
   </PropertyGroup>
 
   <ItemGroup>
