@@ -5,54 +5,46 @@
 using System;
 using System.Threading;
 using Iot.Device.Adc;
+using force_sensitive_resistor;
 
-namespace force_sensitive_resistor
+Console.WriteLine("Hello Fsr408 capacitor Sample!");
+// Use this sample when using ADC for reading          
+StartReadingWithADC();
+
+// Use this sample if using capacitor for reading
+// StartReadingWithCapacitor();
+
+void StartReadingWithADC()
 {
-    class Program
+    FsrWithAdcSample fsrWithAdc = new();
+    
+    while (true)
     {
-        static void Main(string[] args)
-        {
-            Console.WriteLine("Hello Fsr408 capacitor Sample!");
-            // Use this sample when using ADC for reading          
-            StartReadingWithADC();
+        int value = fsrWithAdc.Read(0);
+        double voltage = fsrWithAdc.CalculateVoltage(value);
+        double resistance = fsrWithAdc.CalculateFsrResistance(voltage);
+        double force = fsrWithAdc.CalculateForce(resistance);
+        Console.WriteLine($"Read value: {value}, milli voltage: {voltage.ToString("f2")}, resistance: {resistance.ToString("f2")}, approximate force in Newtons: {force.ToString("f2")}");
+        Thread.Sleep(500);
+    }
+}
 
-            // Use this sample if using capacitor for reading
-            // StartReadingWithCapacitor();
+void StartReadingWithCapacitor()
+{
+    FsrWithCapacitorSample fsrWithCapacitor = new();
+
+    while (true)
+    {
+        int value = fsrWithCapacitor.ReadCapacitorChargingDuration();
+
+        if (value == 30000)
+        {   // 30000 is count limit, if we got this count it means Fsr has its highest resistance, so it is not pressed
+            Console.WriteLine("Not pressed");
         }
-
-        private static void StartReadingWithADC()
+        else
         {
-            FsrWithAdcSample fsrWithAdc = new FsrWithAdcSample();
-            
-            while (true)
-            {
-                int value = fsrWithAdc.Read(0);
-                double voltage = fsrWithAdc.CalculateVoltage(value);
-                double resistance = fsrWithAdc.CalculateFsrResistance(voltage);
-                double force = fsrWithAdc.CalculateForce(resistance);
-                Console.WriteLine($"Read value: {value}, milli voltage: {voltage.ToString("f2")}, resistance: {resistance.ToString("f2")}, approximate force in Newtons: {force.ToString("f2")}");
-                Thread.Sleep(500);
-            }
+            Console.WriteLine($"Pressed {value}");
         }
-
-        private static void StartReadingWithCapacitor()
-        {
-            FsrWithCapacitorSample fsrWithCapacitor = new FsrWithCapacitorSample();
-
-            while (true)
-            {
-                int value = fsrWithCapacitor.ReadCapacitorChargingDuration();
-
-                if (value == 30000)
-                {   // 30000 is count limit, if we got this count it means Fsr has its highest resistance, so it is not pressed
-                    Console.WriteLine("Not pressed");
-                }
-                else
-                {
-                    Console.WriteLine($"Pressed {value}");
-                }
-                Thread.Sleep(500);
-            }
-        }
+        Thread.Sleep(500);
     }
 }
