@@ -10,6 +10,8 @@ using System.Runtime.InteropServices;
 using System.Threading;
 using System.Threading.Tasks;
 
+#pragma warning disable SA1011 // Closing square brackets should be spaced correctly https://github.com/DotNetAnalyzers/StyleCopAnalyzers/issues/2927
+
 namespace System.Device.Gpio.Drivers
 {
     /// <summary>
@@ -29,11 +31,11 @@ namespace System.Device.Gpio.Drivers
         private const string DeviceTreeRanges = "/proc/device-tree/soc/ranges";
         private const string ModelFilePath = "/proc/device-tree/model";
 
-        private readonly PinState[] _pinModes;
+        private readonly PinState?[] _pinModes;
         private RegisterView* _registerViewPointer = null;
         private static readonly object s_initializationLock = new object();
 
-        private UnixDriver _interruptDriver = null;
+        private UnixDriver? _interruptDriver = null;
 
         /// <summary>
         /// Returns true if this is a Raspberry Pi4
@@ -113,9 +115,8 @@ namespace System.Device.Gpio.Drivers
         {
             ValidatePinNumber(pinNumber);
 
-            _interruptDriver.OpenPin(pinNumber);
-            _pinModes[pinNumber].InUseByInterruptDriver = true;
-
+            _interruptDriver!.OpenPin(pinNumber);
+            _pinModes[pinNumber]!.InUseByInterruptDriver = true;
             _interruptDriver.AddCallbackForPinValueChangedEvent(pinNumber, eventTypes, callback);
         }
 
@@ -127,10 +128,9 @@ namespace System.Device.Gpio.Drivers
         {
             ValidatePinNumber(pinNumber);
 
-            bool isOpen = _pinModes[pinNumber] != null && _pinModes[pinNumber].InUseByInterruptDriver;
-            if (isOpen)
+            if (_pinModes[pinNumber]?.InUseByInterruptDriver ?? false)
             {
-                _interruptDriver.ClosePin(pinNumber);
+                _interruptDriver!.ClosePin(pinNumber);
             }
 
             // Set pin low and mode to input upon closing a pin
@@ -197,8 +197,8 @@ namespace System.Device.Gpio.Drivers
         {
             ValidatePinNumber(pinNumber);
 
-            _interruptDriver.OpenPin(pinNumber);
-            _pinModes[pinNumber].InUseByInterruptDriver = true;
+            _interruptDriver!.OpenPin(pinNumber);
+            _pinModes[pinNumber]!.InUseByInterruptDriver = true;
 
             _interruptDriver.RemoveCallbackForPinValueChangedEvent(pinNumber, callback);
         }
@@ -234,9 +234,9 @@ namespace System.Device.Gpio.Drivers
             register |= (mode == PinMode.Output ? 1u : 0u) << shift;
             *registerPointer = register;
 
-            if (_pinModes[pinNumber] != null)
+            if (_pinModes[pinNumber] is object)
             {
-                _pinModes[pinNumber].CurrentPinMode = mode;
+                _pinModes[pinNumber]!.CurrentPinMode = mode;
             }
             else
             {
@@ -378,8 +378,8 @@ namespace System.Device.Gpio.Drivers
         {
             ValidatePinNumber(pinNumber);
 
-            _interruptDriver.OpenPin(pinNumber);
-            _pinModes[pinNumber].InUseByInterruptDriver = true;
+            _interruptDriver!.OpenPin(pinNumber);
+            _pinModes[pinNumber]!.InUseByInterruptDriver = true;
 
             return _interruptDriver.WaitForEvent(pinNumber, eventTypes, cancellationToken);
         }
@@ -395,8 +395,8 @@ namespace System.Device.Gpio.Drivers
         {
             ValidatePinNumber(pinNumber);
 
-            _interruptDriver.OpenPin(pinNumber);
-            _pinModes[pinNumber].InUseByInterruptDriver = true;
+            _interruptDriver!.OpenPin(pinNumber);
+            _pinModes[pinNumber]!.InUseByInterruptDriver = true;
 
             return _interruptDriver.WaitForEventAsync(pinNumber, eventTypes, cancellationToken);
         }
@@ -623,11 +623,8 @@ namespace System.Device.Gpio.Drivers
                 _registerViewPointer = null;
             }
 
-            if (_interruptDriver != null)
-            {
-                _interruptDriver.Dispose();
-                _interruptDriver = null;
-            }
+            _interruptDriver?.Dispose();
+            _interruptDriver = null;
         }
 
         private class PinState
