@@ -96,23 +96,17 @@ namespace System.Device.Gpio
         private RaspberryBoardInfo(Dictionary<string, string> settings)
         {
             _settings = settings;
-            Firmware = 0;
-            ProcessorName = string.Empty;
 
-            if (_settings.TryGetValue("Hardware", out string hardware))
-            {
-                ProcessorName = hardware;
-            }
+            ProcessorName = _settings.TryGetValue("Hardware", out string? hardware) && hardware is object ? hardware : string.Empty;
 
-            if (_settings.TryGetValue("Revision", out string revision)
+            if (_settings.TryGetValue("Revision", out string? revision)
                 && !string.IsNullOrEmpty(revision)
                 && int.TryParse(revision, NumberStyles.HexNumber, CultureInfo.InvariantCulture, out int firmware))
             {
                 Firmware = firmware;
             }
 
-            if (_settings.TryGetValue("Serial", out string serial)
-                && !string.IsNullOrEmpty(serial))
+            if (_settings.TryGetValue("Serial", out string? serial))
             {
                 SerialNumber = serial;
             }
@@ -215,7 +209,7 @@ namespace System.Device.Gpio
         /// <summary>
         /// Gets the serial number.
         /// </summary>
-        public string SerialNumber
+        public string? SerialNumber
         {
             get;
         }
@@ -230,8 +224,7 @@ namespace System.Device.Gpio
         {
             get
             {
-                var firmware = Firmware;
-                return (firmware & 0xFFFF0000) != 0;
+                return (Firmware & 0xFFFF0000) != 0;
             }
         }
 
