@@ -2,14 +2,12 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-using System.Runtime.CompilerServices;
-
 namespace System.Device.I2c
 {
     /// <summary>
     /// The communications channel to a device on an I2C bus.
     /// </summary>
-    public abstract class I2cDevice : IDisposable
+    public abstract partial class I2cDevice : IDisposable
     {
         /// <summary>
         /// The connection settings of a device on an I2C bus. The connection settings are immutable after the device is created
@@ -75,16 +73,6 @@ namespace System.Device.I2c
             {
                 return new UnixI2cDevice(settings);
             }
-        }
-
-        [MethodImpl(MethodImplOptions.NoInlining)]
-        private static I2cDevice CreateWindows10I2cDevice(I2cConnectionSettings settings)
-        {
-            // This wrapper is needed to prevent Mono from loading Windows10I2cDevice
-            // which causes all fields to be loaded - one of such fields is WinRT type which does not
-            // exist on Linux which causes TypeLoadException.
-            // Using NoInlining and no explicit type prevents this from happening.
-            return new Windows10I2cDevice(settings);
         }
 
         /// <inheritdoc cref="IDisposable.Dispose"/>
